Load the images of the objects at game's setup
The reason to do it is to avoid reading from disk the image of the object every
time it is created since it is time consuming. This works better for objects
that are constantly created, for example like the bullets.

With this new implementation the image is already in RAM and the object only has
to copy it from there.